requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
python-decouple==3.8
python-ulid==1.1.0
pytz==2023.3
Werkzeug==2.3.6
